api/nlp2sql_trainer/requirements.txt
oracledb>=2.1.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
sentence-transformers>=3.0.1
#faiss-cpu>=1.8.0
nltk>=3.9
pyyaml>=6.0.2
click>=8.1.7
tqdm>=4.66.4
tabulate>=0.9.0
streamlit>=1.36.0
pyarrow>=16.1.0
